refactor(#105): 앨범 업로드 기능 앨범서비스에 통합
